+++ src/Snake.cpp
@@ -1,24 +1,9 @@
 // Snake.cpp
 #include <iostream>
 #include <SFML\Graphics.hpp>
+#include "GeneralFunctions.h"
 #include "Snake.h"
 #include "SnakeFood.h"
-
-// Test for collision between two objects
-bool SnakeCollision(sf::RectangleShape object1, sf::RectangleShape object2)
-{
-    // Hack for getting a temporary collision
-    // Collision only tested for origin corrordinate
-    sf::Vector2f object1Position = object1.getPosition();
-    sf::Vector2f object2Position = object2.getPosition();
-    if (object1Position.x != object2Position.x)
-        return 0;
-    if (object1Position.y != object2Position.y)
-        return 0;
-
-    return 1;
-
-}
 
 // Check keyboard for new direction of snake
 void Snake::CheckDirection()
@@ -49,22 +34,6 @@ bool Snake::CheckBoundaries()
     return false;
 }
 
-// Get a new coordinate position based on snake direction
-sf::Vector2f CalculateNewPosition(int direction, sf::Vector2f position)
-{
-    if (direction == 0)
-        return position;
-    if (direction == 1)
-        position.x -= 25;
-    if (direction == 2)
-        position.y -= 25;
-    if (direction == 3)
-        position.y += 25;
-    if (direction == 4)
-        position.x += 25;
-    return position;
-}
-
 // Move snake based on direction and test for eating food
 void Snake::MoveSnake(SnakeFood& snakeFood, sf::VideoMode gameVideoMode)
 {
@@ -82,7 +51,7 @@ void Snake::MoveSnake(SnakeFood& snakeFood, sf::VideoMode gameVideoMode)
     }
     newBodyPart.setFillColor(sf::Color::Green);
     snakeBody.push_front(newBodyPart);
-    if (!SnakeCollision(GetSnakeHead(), snakeFood.snakeFoodObject))
+    if (!GlobalCollision(GetSnakeHead().getPosition(), snakeFood.snakeFoodObject.getPosition()))
         snakeBody.pop_back();
     SnakeFoodCollision(snakeFood, gameVideoMode);
     return;
@@ -119,7 +88,7 @@ bool Snake::IsSelfCollision(sf::RectangleShape testRectangle)
 {
     for (auto it = snakeBody.cbegin(); it != snakeBody.cend(); ++it)
     {
-        if (SnakeCollision(testRectangle, *it))
+        if (GlobalCollision(testRectangle.getPosition(), (*it).getPosition()))
         {
             return true;
         }
